@@ -5,13 +5,15 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+
+import com.google.android.material.navigation.NavigationView;
+
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -251,9 +253,9 @@ public class DashboardActivity extends AppCompatActivity implements NavigationVi
         ViewStub stub = findViewById(R.id.dashboard_layout_loader);
 
         if (sessionApp.getTYPEUSER().equals(User.UserTypes.PARENTUSER))
-            stub.setLayoutResource(R.layout.activity_dashboard_parent_content);
+            stub.setLayoutResource(R.layout.activity_dashboard__parent_content);
         if (sessionApp.getTYPEUSER().equals(User.UserTypes.TEACHERUSER))
-            stub.setLayoutResource(R.layout.activity_dashboard_teacher_content);
+            stub.setLayoutResource(R.layout.activity_dashboard__teacher_content);
         stub.inflate();
     }
 
