@@ -3,7 +3,8 @@ package com.workingtogether.workingtogether.activities;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 
 import com.workingtogether.workingtogether.R;
