@@ -1,7 +1,8 @@
 package com.workingtogether.workingtogether.activities.teacher;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.workingtogether.workingtogether.R;
 
