@@ -2,7 +2,8 @@ package com.workingtogether.workingtogether.util;
 
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
+
+import androidx.appcompat.app.AlertDialog;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
@@ -12,10 +13,9 @@ public class Util {
     public static class Dates {
 
         public static String getDateTime() {
-            Calendar c = Calendar.getInstance();
+            Calendar calendar = Calendar.getInstance();
             SimpleDateFormat dateFormatormat = new SimpleDateFormat("dd/MM/yyyy h:mm a");
-            String dateTime = dateFormatormat.format(c.getTime());
-            return dateTime;
+            return dateFormatormat.format(calendar.getTime());
         }
 
     }
